+++ src/main/java/com/funcxy/oj/controllers/UserController.java
@@ -216,86 +216,4 @@ public class UserController {
                 ,HttpStatus.OK);
     }
 
-    //收藏题单
-    @RequestMapping(value = "/{username}/liked-problem-lists/{id}",method = RequestMethod.POST)
-    public ResponseEntity<Object> likeProblemLists(@PathVariable String username,@PathVariable ObjectId problemListId,HttpSession httpSession){
-        if (UserUtil.isSignedIn(httpSession)){
-            User user = userRepository.findById((ObjectId)httpSession.getAttribute("userId"));
-            if (user == null) return new ResponseEntity<Object>(new NotFoundError(),HttpStatus.NOT_FOUND);
-            if (userRepository.findById((ObjectId) httpSession.getAttribute("userId")).getProblemListLiked().indexOf(problemListId)!=-1)return new ResponseEntity<Object>(new FieldsDuplicateError(),HttpStatus.BAD_REQUEST);
-            userRepository.findById((ObjectId) httpSession.getAttribute("userId")).addProblemListLiked(problemListId);
-            return new ResponseEntity<Object>(new User(),HttpStatus.OK);
-        }else {
-            return new ResponseEntity<Object>(new ForbiddenError(),HttpStatus.FORBIDDEN);
-        }
-    }
-
-    //取消收藏题单
-    @RequestMapping(value = "/{username}/liked-problem-lists/{id}",method = RequestMethod.DELETE)
-    public ResponseEntity<Object> dislikeProblemLists(@PathVariable String username,@PathVariable ObjectId problemListId,HttpSession httpSession){
-        if (UserUtil.isSignedIn(httpSession)){
-            User user = userRepository.findById((ObjectId)httpSession.getAttribute("userId"));
-            if (user == null) return new ResponseEntity<Object>(new NotFoundError(),HttpStatus.NOT_FOUND);
-            if (userRepository.findById((ObjectId) httpSession.getAttribute("userId")).getProblemListLiked().indexOf(problemListId)==-1)return new ResponseEntity<Object>(new NotFoundError(),HttpStatus.NOT_FOUND);
-            userRepository.findById((ObjectId) httpSession.getAttribute("userId")).deleteProblemListLiked(problemListId);
-            return new ResponseEntity<Object>(new User(),HttpStatus.OK);
-        }else {
-            return new ResponseEntity<Object>(new ForbiddenError(),HttpStatus.FORBIDDEN);
-        }
-    }
-    //获取收藏的题单
-    @RequestMapping(value = "/{username}/liked-problem-lists",method = RequestMethod.GET)
-    public ResponseEntity getLikedProblemLists(@PathVariable String username,HttpSession httpSession){
-        if (UserUtil.isSignedIn(httpSession)){
-            User user = userRepository.findById((ObjectId)httpSession.getAttribute("userId"));
-            if (user == null) return  new ResponseEntity<>(new NotFoundError(),HttpStatus.NOT_FOUND);
-            List<ObjectId> likedProblemList = user.getProblemListLiked();
-            return new ResponseEntity<>(new PageImpl<CleanedProblemList>(
-                    likedProblemList
-                            .stream()
-                            .map(
-                                pro->new CleanedProblemList(
-                                        problemListRepository.findById(pro).getId(),
-                                        problemListRepository.findById(pro).getTitle(),
-                                        problemListRepository.findById(pro).getType()
-                                )
-                            )
-                            .collect(Collectors.toList())
-                    ,pageable
-                    ,likedProblemList.size()
-            )
-                    ,HttpStatus.FOUND
-            );
-        }else {
-            return new ResponseEntity<>(new ForbiddenError(),HttpStatus.FORBIDDEN);
-        }
-    }
-    //获取收藏的题目
-    @RequestMapping(value = "/{username}/liked-problems",method = RequestMethod.GET)
-    public ResponseEntity getLikedProblems(@PathVariable String username,HttpSession httpSession){
-        if (!UserUtil.isSignedIn(httpSession)){
-            return new ResponseEntity<>(new ForbiddenError(),HttpStatus.FORBIDDEN);
-        }
-        User user = userRepository.findById((ObjectId) httpSession.getAttribute("userId"));
-        if (user==null){
-            return new ResponseEntity<>(new NotFoundError(),HttpStatus.NOT_FOUND);
-        }
-        List<ObjectId> likedProblem = user.getProblemLiked();
-
-        return new ResponseEntity<>(
-                new PageImpl<CleanedProblem>(
-                        likedProblem.stream()
-                        .map(
-                                pro->new CleanedProblem(
-                                        problemRepository.findById(pro).getId(),
-                                        problemRepository.findById(pro).getTitle()
-                                )
-                        ).collect(Collectors.toList()),
-                        pageable,
-                        likedProblem.size()
-                ),
-                HttpStatus.OK
-        );
-    }
-
 }
